✨ Initial implementation for migrating users
Note that this implementation is not secure:

We do not verify that the newly signed in user was in fact the user they want to upgrade data from. In a worst case scenario, an attacker might call the function using an anonymous UID and request updating data from another user, thus hijacking all their data.

A better solution requires sending ID tokens and verifying them on the backend before making any changes.